[MA-6] refactor : 패키지명 변경, onboardingPhoneAUthorization 수정
