+++ presentation/src/main/java/com/motax/modutaxi/presentation/ui/intro/signup/identification/OnboardingIdentificationFragment.kt
@@ -29,7 +29,7 @@ class OnboardingIdentificationFragment :
 
     private fun NavController.toQuestion() {
         val action =
-            OnboardingIdentificationFragmentDirections.actionIdentificationFragmentToQuestionFragment()
+            OnboardingIdentificationFragmentDirections.actionOnboardingIdentificationFragmentToOnboardingPhoneAuthorization2()
         navigate(action)
     }
 }
