+++ presentation/src/main/java/com/motax/modutaxi/presentation/ui/intro/signup/phoneauth/OnboardingPhoneAuthorization.kt
@@ -1,0 +1,15 @@
+package com.motax.modutaxi.presentation.ui.intro.signup.phoneauth
+
+import android.os.Bundle
+import android.view.View
+import com.motax.modutaxi.presentation.R
+import com.motax.modutaxi.presentation.base.BaseFragment
+import com.motax.modutaxi.presentation.databinding.FragmentOnboardingPhoneAuthorizationBinding
+
+class OnboardingPhoneAuthorization: BaseFragment<FragmentOnboardingPhoneAuthorizationBinding>(R.layout.fragment_onboarding_phone_authorization) {
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+    }
+}
