+++ presentation/src/main/res/layout/fragment_onboarding_phone_authorization.xml
@@ -70,8 +70,7 @@
             android:id="@+id/tv_identification_description"
             android:layout_width="0dp"
             android:layout_height="wrap_content"
-            android:layout_marginLeft="8dp"
-            android:text="적어주신 번호로\n 인증번호가 전송됐어요!"
+            android:text="적어주신 번호로\n인증번호가 전송됐어요!"
             android:textColor="#000000"
             android:textSize="24sp"
             app:layout_constraintStart_toEndOf="@id/guide_08_vertical"
@@ -108,10 +107,12 @@
                 android:hint="인증번호를 입력해주세요"
                 android:textColorHint="#C0C0C0"
                 android:textSize="16sp"
+                app:layout_constraintEnd_toStartOf="@id/tv_time"
                 app:layout_constraintStart_toStartOf="parent"
                 app:layout_constraintTop_toBottomOf="@id/tv_name_label" />
 
             <TextClock
+                android:id="@+id/tv_time"
                 android:layout_width="0dp"
                 android:layout_height="wrap_content"
                 android:layout_marginTop="6dp"
